+++ src/audio/dai.c
@@ -8,12 +8,9 @@
 #include <sof/audio/buffer.h>
 #include <sof/audio/component_ext.h>
 #include <sof/audio/format.h>
-#include <sof/audio/pcm_converter.h>
 #include <sof/audio/pipeline.h>
-#include <sof/audio/ipc-config.h>
 #include <sof/common.h>
 #include <sof/debug/panic.h>
-#include <sof/drivers/edma.h>
 #include <sof/drivers/interrupt.h>
 #include <sof/ipc/msg.h>
 #include <sof/drivers/interrupt.h>
@@ -21,7 +18,6 @@
 #include <sof/lib/alloc.h>
 #include <sof/lib/cache.h>
 #include <sof/lib/dai.h>
-#include <sof/lib/dma.h>
 #include <sof/lib/memory.h>
 #include <sof/lib/notifier.h>
 #include <sof/lib/uuid.h>
@@ -45,38 +41,10 @@ DECLARE_SOF_RT_UUID("dai", dai_comp_uuid, 0xc2b00d27, 0xffbc, 0x4150,
 
 DECLARE_TR_CTX(dai_comp_tr, SOF_UUID(dai_comp_uuid), LOG_LEVEL_INFO);
 
-struct dai_data {
-	/* local DMA config */
-	struct dma_chan_data *chan;
-	uint32_t stream_id;
-	struct dma_sg_config config;
-	struct comp_buffer *dma_buffer;
-	struct comp_buffer *local_buffer;
-	struct timestamp_cfg ts_config;
-	struct dai *dai;
-	struct dma *dma;
-	struct dai_group *group;	/**< NULL if no group assigned */
-	int xrun;		/* true if we are doing xrun recovery */
-
-	pcm_converter_func process;	/* processing function */
-
-	uint32_t dai_pos_blks;	/* position in bytes (nearest block) */
-	uint64_t start_position;	/* position on start */
-	uint32_t period_bytes;	/**< number of bytes per one period */
-
-	/* host can read back this value without IPC */
-	uint64_t *dai_pos;
-
-	struct sof_ipc_dai_config *dai_config;	/* dai_config from the host */
-	struct ipc_config_dai ipc_config;	/* base config at DAI new */
-
-	uint64_t wallclock;	/* wall clock at stream start */
-};
-
 static void dai_atomic_trigger(void *arg, enum notify_id type, void *data);
 
 /* Assign DAI to a group */
-static int dai_assign_group(struct comp_dev *dev, uint32_t group_id)
+int dai_assign_group(struct comp_dev *dev, uint32_t group_id)
 {
 	struct dai_data *dd = comp_get_drvdata(dev);
 
@@ -257,8 +225,8 @@ static void dai_free(struct comp_dev *dev)
 
 	dai_put(dd->dai);
 
-	if (dd->dai_config)
-		rfree(dd->dai_config);
+	if (dd->dai_spec_config)
+		rfree(dd->dai_spec_config);
 
 	rfree(dd);
 	rfree(dev);
@@ -339,73 +307,6 @@ static int dai_verify_params(struct comp_dev *dev,
 	/* set component period frames */
 	component_set_period_frames(dev, params->rate);
 
-	return 0;
-}
-
-static int dai_data_config(struct comp_dev *dev)
-{
-	struct dai_data *dd = comp_get_drvdata(dev);
-	struct ipc_config_dai *dai = &dd->ipc_config;
-
-	if (!dd->dai_config) {
-		comp_err(dev, "dai_data_config(): no config set for dai %d type %d",
-			 dai->dai_index, dai->type);
-		return -EINVAL;
-	}
-
-	comp_info(dev, "dai_data_config() dai type = %d index = %d dd %p",
-		  dai->type, dai->dai_index, dd);
-
-	/* cannot configure DAI while active */
-	if (dev->state == COMP_STATE_ACTIVE) {
-		comp_info(dev, "dai_data_config(): Component is in active state.");
-		return 0;
-	}
-
-	switch (dd->dai_config->type) {
-	case SOF_DAI_INTEL_SSP:
-		/* set dma burst elems to slot number */
-		dd->config.burst_elems = dd->dai_config->ssp.tdm_slots;
-		break;
-	case SOF_DAI_INTEL_DMIC:
-		/* We can use always the largest burst length. */
-		dd->config.burst_elems = 8;
-
-		comp_info(dev, "config->dmic.fifo_bits = %u config->dmic.num_pdm_active = %u",
-			  dd->dai_config->dmic.fifo_bits,
-			  dd->dai_config->dmic.num_pdm_active);
-		break;
-	case SOF_DAI_INTEL_HDA:
-		break;
-	case SOF_DAI_INTEL_ALH:
-		/* SDW HW FIFO always requires 32bit MSB aligned sample data for
-		 * all formats, such as 8/16/24/32 bits.
-		 */
-		dev->ipc_config.frame_fmt = SOF_IPC_FRAME_S32_LE;
-		dd->dma_buffer->stream.frame_fmt = dev->ipc_config.frame_fmt;
-
-		dd->config.burst_elems =
-			dd->dai->plat_data.fifo[dai->direction].depth;
-
-		/* As with HDA, the DMA channel is assigned in runtime,
-		 * not during topology parsing.
-		 */
-		dd->stream_id = dd->dai_config->alh.stream_id;
-		break;
-	case SOF_DAI_IMX_SAI:
-		COMPILER_FALLTHROUGH;
-	case SOF_DAI_IMX_ESAI:
-		dd->config.burst_elems =
-			dd->dai->plat_data.fifo[dai->direction].depth;
-		break;
-	default:
-		/* other types of DAIs not handled for now */
-		comp_warn(dev, "dai_data_config(): Unknown dai type %d",
-			  dd->dai_config->type);
-		break;
-	}
-
-	/* some DAIs may not need extra config */
 	return 0;
 }
 
@@ -549,7 +450,7 @@ static int dai_params(struct comp_dev *dev,
 	comp_dbg(dev, "dai_params()");
 
 	/* configure dai_data first */
-	err = dai_data_config(dev);
+	err = ipc_dai_data_config(dev);
 	if (err < 0)
 		return err;
 
@@ -659,102 +560,6 @@ static int dai_params(struct comp_dev *dev,
 		dai_capture_params(dev, period_bytes, period_count);
 }
 
-static int dai_config_dma_channel(struct comp_dev *dev, struct sof_ipc_dai_config *config)
-{
-	struct dai_data *dd = comp_get_drvdata(dev);
-	struct ipc_config_dai *dai = &dd->ipc_config;
-	int channel;
-	int handshake;
-
-	assert(config);
-
-	switch (config->type) {
-	case SOF_DAI_INTEL_SSP:
-		COMPILER_FALLTHROUGH;
-	case SOF_DAI_INTEL_DMIC:
-		channel = 0;
-		break;
-	case SOF_DAI_INTEL_HDA:
-		channel = config->hda.link_dma_ch;
-		break;
-	case SOF_DAI_INTEL_ALH:
-		/* As with HDA, the DMA channel is assigned in runtime,
-		 * not during topology parsing.
-		 */
-		channel = config->alh.stream_id;
-		break;
-	case SOF_DAI_IMX_SAI:
-		COMPILER_FALLTHROUGH;
-	case SOF_DAI_IMX_ESAI:
-		handshake = dai_get_handshake(dd->dai, dai->direction,
-					      dd->stream_id);
-		channel = EDMA_HS_GET_CHAN(handshake);
-		break;
-	default:
-		/* other types of DAIs not handled for now */
-		comp_err(dev, "dai_config_dma_channel(): Unknown dai type %d",
-			 config->type);
-		channel = DMA_CHAN_INVALID;
-		break;
-	}
-
-	return channel;
-}
-
-static int dai_config(struct comp_dev *dev, struct sof_ipc_dai_config *config)
-{
-	struct dai_data *dd = comp_get_drvdata(dev);
-	int ret;
-
-	comp_info(dev, "dai_config() check dai type = %d index = %d dd %p",
-			  config->type, config->dai_index, dd);
-
-	/* ignore if message not for this DAI id/type */
-	if (dd->ipc_config.dai_index != config->dai_index ||
-	    dd->ipc_config.type != config->type)
-		return 0;
-
-	comp_info(dev, "dai_config() dai type = %d index = %d dd %p",
-		  config->type, config->dai_index, dd);
-
-	/* cannot configure DAI while active */
-	if (dev->state == COMP_STATE_ACTIVE) {
-		comp_info(dev, "dai_config(): Component is in active state. Ignore config");
-		return 0;
-	}
-
-	if (dd->chan) {
-		comp_info(dev, "dai_config(): Configured. dma channel index %d, ignore...",
-			  dd->chan->index);
-		return 0;
-	}
-
-	if (config->group_id) {
-		ret = dai_assign_group(dev, config->group_id);
-
-		if (ret)
-			return ret;
-	}
-
-	/* do nothing for asking for channel free, for compatibility. */
-	if (dai_config_dma_channel(dev, config) == DMA_CHAN_INVALID)
-		return 0;
-
-	/* allocated dai_config if not yet */
-	if (!dd->dai_config) {
-		dd->dai_config = rzalloc(SOF_MEM_ZONE_RUNTIME, 0, SOF_MEM_CAPS_RAM,
-					 sizeof(*dd->dai_config));
-		if (!dd->dai_config) {
-			comp_err(dev, "dai_config(): No memory for dai_config.");
-			return -ENOMEM;
-		}
-	}
-
-	*dd->dai_config = *config;
-
-	return 0;
-}
-
 static int dai_config_prepare(struct comp_dev *dev)
 {
 	struct dai_data *dd = comp_get_drvdata(dev);
@@ -766,8 +571,8 @@ static int dai_config_prepare(struct comp_dev *dev)
 		return 0;
 	}
 
-	if (!dd->dai_config) {
-		comp_err(dev, "dai_config is not set yet!");
+	if (!dd->dai_spec_config) {
+		comp_err(dev, "dai specific config is not set yet!");
 		return -EINVAL;
 	}
 
@@ -777,7 +582,7 @@ static int dai_config_prepare(struct comp_dev *dev)
 		return 0;
 	}
 
-	channel = dai_config_dma_channel(dev, dd->dai_config);
+	channel = dai_config_dma_channel(dev, dd->dai_spec_config);
 	comp_info(dev, "dai_config_prepare(), channel = %d", channel);
 
 	/* do nothing for asking for channel free, for compatibility. */
