+++ zephyr/CMakeLists.txt
@@ -354,6 +354,7 @@ zephyr_library_sources(
 	${SOF_IPC_PATH}/ipc-common.c
 	${SOF_IPC_PATH}/handler-ipc3.c
 	${SOF_IPC_PATH}/helper-ipc3.c
+	${SOF_IPC_PATH}/dai-ipc3.c
 	${SOF_IPC_PATH}/ipc-host-ptable.c
 	${SOF_SRC_PATH}/spinlock.c
